refactor: supabaseで自動生成されるファイルの名前を変更
