+++ api/client.ts
@@ -1,5 +1,5 @@
 import { createClient } from "@supabase/supabase-js"
-import { Database } from "../types/schema"
+import { Database } from "../types/supabase"
 
 const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL ?? ""
 const SUPABASE_KEY = process.env.NEXT_PUBLIC_SUPABASE_KEY ?? ""
